Complete Code for the MVP
Implemented the complete code for the startup mvp, where the user can get register, the driver can delivery food depending, restaurant up to their task also

+++ Services/BasketService/BasketService.cs
@@ -46,45 +46,46 @@ namespace startup_trial.Services.BasketService
                     Dish = dish,
                     Count = 1,
                     User = currentUser,
+                    RestaurantId = dish.RestaurantId
                 };
                 _context.Baskets.Add(addToCart);
             }
-
+            
             await _context.SaveChangesAsync();
         }
 
 
         public async Task DeleteDishInCart(Guid dishId, bool increase, string userId)
         {
+            //var checkUser = restaurants.
             var currentUser = await _userManager.FindByIdAsync(userId);
 
-            if (currentUser == null)
-            {
-                throw new Exception("No active user");
-            }
-
             var dish = await _context.Dishes.FirstOrDefaultAsync(dish => dish.Id == dishId);
-
-            if (dish == null)
+            if (currentUser == null || dish == null)
             {
-                throw new Exception("Dish doesn't exist");
+                if (currentUser is null) { throw new Exception("User is not active"); }
+                else if((dish == null)) { throw new Exception("Dish not found");  }
+                else { throw new Exception("neither dish nor user found"); }
             }
 
-            var dishInCart = await _context.Baskets.FirstOrDefaultAsync(dishInCart => dishInCart.Dish.Id == dishId && dishInCart.User.Id == currentUser.Id && dishInCart.Order.Id == null);
-
-            if (dishInCart != null)
+            if (currentUser is Customer customer)
             {
-                if (increase == true)
-                {
-                    dishInCart.Count -= 1;
-                }
+                var dishInCart = await _context.Baskets.FirstOrDefaultAsync(dishInCart => dishInCart.Dish.Id == dishId && dishInCart.User.Id == currentUser.Id && dishInCart.Order.Id == null);
 
-                else
+                if (dishInCart != null)
                 {
-                    _context.Baskets.Remove(dishInCart);
+                    if (increase == true)
+                    {
+                        dishInCart.Count -= 1;
+                    }
+
+                    else
+                    {
+                        _context.Baskets.Remove(dishInCart);
+                    }
                 }
+                await _context.SaveChangesAsync();
             }
-            await _context.SaveChangesAsync();
         }
 
 
@@ -99,23 +100,31 @@ namespace startup_trial.Services.BasketService
                 return new List<DishBasketDto>();
             }
 
-            var dishInCartList = await _context.Baskets.Where(basket => basket.User.Id == currentUser.Id && basket.Order.Id == null).Include(basket => basket.Dish).ToListAsync();
-
-
-            var cartList = dishInCartList.Select(basket => new DishBasketDto
+            if (currentUser is Customer customer)
             {
-                Name = basket.Dish.Name,
+                var dishInCartList = await _context.Baskets.Where(basket => basket.User.Id == currentUser.Id && basket.Order.Id == null).Include(basket => basket.Dish).ToListAsync();
 
-                Price = basket.Dish.Price,
 
-                TotalPrice = basket.Dish.Price * basket.Count,
+                var cartList = dishInCartList.Select(basket => new DishBasketDto
+                {
+                    Name = basket.Dish.Name,
 
-                Amount = basket.Count,
+                    Price = basket.Dish.Price,
 
-                Image = basket.Dish.PhotoUrl,
-            }).ToList();
+                    TotalPrice = basket.Dish.Price * basket.Count,
 
-            return cartList;
+                    Amount = basket.Count,
+
+                    Image = basket.Dish.PhotoUrl,
+                }).ToList();
+
+                return cartList;
+            }
+            
+            else
+            {
+                throw new Exception("User don't have authorization for this task");
+            }
         }
     }
 }
